pkg/config: Support slackID field
This new slackID field will be used to ping users on Slack. Team manager
only needs to know about it to allow it in the configuration file; it's
not used by Team manager itself. All Slack user IDs must refer to users
in the same Slack workspace, designated by the new slackWorkspace field.

Signed-off-by: Paul Chaignon <paul@cilium.io>