cobra: refactor error handling
Currently, most cobra commands are panic in case of an error without
using the provided facility by cobra to properly propagate errors to the
user.

This commit refactors the code to returning the errors back to the main
by using cobra command.RunE which includes an error as return type of
the commands.

This way, the error automatically gets displayed to the user alongside
the command information.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>