cmd: refactor pull into init command
The command `pull` only pulls the config from GitHub if the local config
doesn't exist. This has its reason, because not all information can be
gathered from GitHub (e.g. exclusions, slack-id, ...).

It is helpful to bootstrap and initialize a local config file.
Therefore, this commit renames the command `pull` to `init` and moves it
into its own file.

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>