cmd: Sort local configuration before applying it
Previously there was an issue where dry run may report that there is a
diff for some team configuration due to the ordering of members in the
Config.AllTeams[...].Members slice. The remote configuration pulled from
GitHub was already sorted, but the locally read configuration was not.
Fix this issue to address spurious reports of out of sync configuration.

As part of this, in addition to sorting the members and mentors for a
team, the sort function must also sort that team's child team. Implement
a simple recursive function to apply this down the tree.

Signed-off-by: Joe Stringer <joe@cilium.io>